+++ RELEASE-NOTES.md
@@ -1,10 +1,11 @@
 ## RELEASE NOTES
 
-### Version 4.7.6-tab-not-defined"
+### Version 4.7.6-tab-not-defined-error"
 **EUI-4562** Case details Active tasks tab on success not returning to active tasks page
 
 ### Version 4.7.6-hearings-tab
 **EUI-4441** Add hearings tab in case details
+**EUI-4627** Can't access /hearings tab from URL
 
 ### Version 4.7.0-EUI-4412-order-fields-CYA-page
 **EUI-4414** Field order on the CYA page changes